Rename http_server.h
