+++ sherpa/cpp_api/websocket/online_websocket_server_impl.h
@@ -32,7 +32,7 @@
 #include "sherpa/cpp_api/online_recognizer.h"
 #include "sherpa/cpp_api/online_stream.h"
 #include "sherpa/cpp_api/parse-options.h"
-#include "sherpa/cpp_api/websocket/http_server.h"
+#include "sherpa/cpp_api/websocket/http-server.h"
 #include "sherpa/cpp_api/websocket/tee_stream.h"
 #include "websocketpp/config/asio_no_tls.hpp"  // TODO(fangjun): support TLS
 #include "websocketpp/server.hpp"
